pyproject.toml
[project]
name = "admet-prediction"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.8"
dependencies = [
    "chemprop==2.2.1",
    "gin-config==0.5.0",
    "ipython==8.12.3",
    "lightgbm==4.6.0",
    "lightning==2.5.5",
    "map4==1.1.3",
    "matplotlib==3.10.7",
    "myst-parser>=3.0.1",
    "numpy==2.3.3",
    "pandas==2.3.3",
    "pillow==11.3.0",
    "pyyaml==6.0.3",
    "ray[data,serve,train,tune]==2.49.2",
    "rdkit==2025.3.6",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "sphinx<7.0.0",
    "sphinx-rtd-theme>=0.5.1",
    "torch==2.8.0",
    "umap-learn>0.5.8",
    "xgboost==3.0.5",
]
